% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{check.list.format}
\alias{check.list.format}
\title{check.list.format}
\usage{
check.list.format(data)
}
\arguments{
\item{data}{Three levels list with [[1]] Datamatrix, [[2]] SamplesMetadata, [[3]] VariableMetadata.}
}
\description{
Check the format of 3 levels list :
[[1]] samples.ID x variables.ID
[[2]] samples.ID x sample.metadata (same order than rows in [[1]])
[[3]] variables.ID x variable.metadata (same order than columns in [[1]])
}
\examples{
check.list.format(Data.list)
}
\keyword{check}
\keyword{list,}

