% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{compare_means_letters}
\alias{compare_means_letters}
\title{Statistical test and letters}
\usage{
compare_means_letters(formula, data, method = "t.test", group.by = NULL)
}
\arguments{
\item{formula}{formula y ~ a with y the response value and a the factor (or group of factor) to test (as in ggpubr::compare_means())}

\item{data}{the input data with a metadata and values}

\item{method}{on of method accepted by ggpubr::compare_means()}

\item{group.by}{the grouping variables to use (optional)}
}
\value{
a data.table with the same structure as input data with statisticals results column added and significant levels letters by groups
}
\description{
Description of the function
}
\examples{
compare_means_letters()
}
\keyword{significance}
