% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{d.t.transform}
\alias{d.t.transform}
\title{Transform a datamatrix and replace zero}
\usage{
d.t.transform(data, ZvalL = F, PercZ = 1, RepZeroL = F, Log2L = T)
}
\arguments{
\item{data}{a data.table with first column as row IDs}

\item{ZvalL}{Should Zero val column be deleted ?}

\item{PercZ}{Prop of zero value above which deletion is made}

\item{RepZeroL}{Should remaining zero be replaced by half of minimum value ?}

\item{Log2L}{Should log2 transform be done}
}
\value{
The resulting data.table only
}
\description{
Do any of the two following function : Replace zero by the minimum value divided by 2 and/or tranform data
with Trans.fun function (by default log2).
}
\examples{
d.t.transform()
}
\keyword{transform}
