% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{data.subset}
\alias{data.subset}
\title{Subset list}
\usage{
data.subset(data, Var.sel = NULL, Samples.sel = NULL)
}
\arguments{
\item{data}{list of three dataframes : [[1]] Datamatrix, [[2]] SamplesMetadata [[3]] VariableMetadata.}

\item{Var.sel}{vector of variable to subset (rownames) from [[3]]}

\item{Samples.sel}{vector of samples to subset (rownames) from [[2]]}
}
\description{
Subset a three level list by variables and/or samples.
}
\examples{
data.subset()
}
\keyword{list}
\keyword{subset,}

