% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{Data_sub}
\alias{Data_sub}
\title{SD_data_sub}
\usage{
Data_sub(Data, col_select, ...)
}
\arguments{
\item{Data}{list with [[1]] Datamatrix [[2]] samples metadata}

\item{Results.path.root}{Name for the results folder}

\item{ropls_param}{Parameters to pass to ropls::opls function}

\item{factor_group}{vector of column(s) number/name of metadata table (in Data[[2]]) to use for grouping corcircles}
}
\description{
Subset a list of three dataframes : [[1]] Datamatrix, [[2]] Samples.Metadata [[3]] Variable.Metadata.
[[1]] With samples ids as rownames and variables ids as column names
[[2]] Metadata of sample in column and samples ids as rownames (same order than rows in [[1]])
[[3]] Variables metadata in column and variable ids as rownames (same order than columns in [[1]])
}
\examples{
SD_pca(Data, Results.path.root = "Default", ropls_param = list(predI = 2, plotL = F), factor_group = NULL)
}
\keyword{pca}

