% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dataframe.transform}
\alias{dataframe.transform}
\title{Transform a datamatrix and replace zero}
\usage{
dataframe.transform(data, ZeroL = T, TransL = F, Trans.fun = log2, ...)
}
\arguments{
\item{data}{a dataframe with value only to transform}

\item{ZeroL}{Should zero be replaced by minimum value divided by 2}

\item{TransL}{Should transformation be made}

\item{Trans.fun}{Which transformation to perform (default log2)}

\item{...}{Argument to pass to Trans.fun}
}
\value{
The resulting datamatrix only
}
\description{
New version available : see d.t.transform. Do any of the two following function : Replace zero by the minimum value divided by 2 and/or tranform data
with Trans.fun function (by default log2).
}
\examples{
dataframe.transform()
}
\keyword{transform}
