% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{datamatrix.opls}
\alias{datamatrix.opls}
\title{Perform OPLS and custom plots}
\usage{
datamatrix.opls(Data, Opls.y, Samples.grp = NULL, Variables.grp = NULL,
  Legend.L = T, colorL = F, LabelsL = F)
}
\arguments{
\item{Data}{Argument description}

\item{Opls.y}{Name of the grouping factor used in OPLS}

\item{Samples.grp}{Vector of samples grouping factor (for plot color only)}

\item{Variables.grp}{Vector of variable grouping factor (for plot color only)}

\item{Legend.L}{Logical to draw legends}

\item{colorL}{Logical to use colors instead of black & white}

\item{LabelsL}{Logical to show labels on plot}
}
\value{
A list with (OPLS, Plot, VIPS)
}
\description{
Perform an OPLS analysis using "ropls" package and plot results if asked (scores, loadings, VIPs & summary).
}
\examples{
datamatrix.opls()
}
\keyword{x1,}
\keyword{x2,}
\keyword{x3}

