% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{datamatrix.pca}
\alias{datamatrix.pca}
\title{Perform PCA and custom plot}
\usage{
datamatrix.pca(Data.list, Samples.grp = NULL, Variables.grp = NULL,
  Legend.L = F, colorL = F, Samp.lab.L = F, Var.lab.L = T)
}
\arguments{
\item{Data.list}{A 3 levels list with the data}

\item{Samples.grp}{Factor name used for grouping samples (affect plot only)}

\item{Variables.grp}{Factor name used for grouping variables (affect plot only)}

\item{Legend.L}{Logical for drawing legends}

\item{colorL}{Logical for using color or greyscale}

\item{Samp.lab.L}{Logical for drawing Sample labels}

\item{Var.lab.L}{Logical for drawing Variable labels}
}
\value{
A list with [1] pca results, [2] plot as grobs.
}
\description{
Perform a PCA analysis on a 3 levels list and create custom plot.
}
\examples{
datamatrix.pca()
}
\keyword{ggplot}
\keyword{pca,}

