% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{datamatrix.transform}
\alias{datamatrix.transform}
\title{Transform a datamatrix and replace zero}
\usage{
datamatrix.transform(data, TransL = F, Trans.fun = log2, ...)
}
\arguments{
\item{data}{Argument description}

\item{data}{Argument description}
}
\value{
The resulting datamatrix only
}
\description{
Replace zero by the minimum value / 2 and if checked, tranform data with Trans.fun function (by default log2).
}
\examples{
datamatrix.transform()
}
\keyword{x1,}
\keyword{x2,}
\keyword{x3}

