% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dlist.class}
\alias{dlist.class}
\title{Check dlist class}
\usage{
dlist.class(dlist)
}
\arguments{
\item{dlist}{Three levels list with [[1]] Datamatrix, [[2]] SamplesMetadata, [[3]] VariableMetadata.}
}
\value{
A table with class summary
}
\description{
Check the class of dlist levels and return a summary table
}
\examples{
dlist.class()
}
\keyword{check,}
\keyword{data.table}
\keyword{list,}
