% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dlist.opls.data}
\alias{dlist.opls.data}
\title{Return a list for plotting ropls results}
\usage{
dlist.opls.data(dlist, ropls.result)
}
\arguments{
\item{dlist}{The dlist used for ropls}

\item{ropls.result}{Results of ropls package or dlist.opls.min function}
}
\value{
A list with data for ggplot
}
\description{
Return a list for plotting ropls results
}
\examples{
dlist.opls.data()
}
\keyword{ropls}

