% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dlist.opls.min}
\alias{dlist.opls.min}
\title{Perform pca, pls(da) or opls(da) with ropls}
\usage{
dlist.opls.min(dlist, opls.y = NULL, min = T, plotL = F, ...)
}
\arguments{
\item{dlist}{Three levels list with [[1]] Datamatrix, [[2]] SamplesMetadata, [[3]] VariableMetadata.}

\item{opls.y}{Name of SampleMetadata column to use as y response (quoted).}

\item{min}{Logical to get full resulting object of minimal.}

\item{plotL}{Logical to draw summary plot as in ropls package.}

\item{...}{Any argument used by ropls::opls function.}
}
\value{
The resulting list of opls function (subsetted if min = TRUE).
}
\description{
Perform multivariate analysis with ropls package and return a minimal object with results
for plots.
}
\examples{
dlist.opls.min()
}
\keyword{3levels.list,}
\keyword{xcmsSet}

