% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dlist.opls}
\alias{dlist.opls}
\title{Perform OPLS and custom plots}
\usage{
dlist.opls(dlist, Opls.y = NA, ..., VIPxlabels = F, VIP.thr = 1,
  ShowPlot = T, plotheme.auto.args = list(Samples.grp = NULL, Variables.grp
  = NULL, geom_ellipse = T, Legend.L = T, colorL = T, geom_path = F, labelsL =
  F))
}
\arguments{
\item{dlist}{list of three dataframes : [[1]] Datamatrix, [[2]] SamplesMetadata [[3]] VariableMetadata.}

\item{Opls.y}{Name of the grouping factor used in OPLS}

\item{...}{Arguments to pass to opls function}

\item{VIPxlabels}{Logical to show VIPs plot x labels}

\item{VIP.thr}{Threshold for VIPs selestion and coloring}

\item{ShowPlot}{Logical to draw plots in graphic device}

\item{plotheme.auto.args}{List of arguments to pass to plotheme.auto}
}
\value{
A list with (OPLS, Plot, VIPS)
}
\description{
Perform an OPLS analysis using "ropls" package and plot results if asked (scores, loadings, VIPs & summary).
}
\examples{
dlist.opls()
}
\keyword{ggplot}
\keyword{opls,}
