% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dlist.pca.old}
\alias{dlist.pca.old}
\title{Perform PCA and custom plot}
\usage{
dlist.pca.old(dlist, Samples.grp = NULL, Variables.grp = NULL,
  Legend.L = F, colorL = F, Samp.lab.L = F, Var.lab.L = T,
  palpha = 0.8, psize = 0.8, ShowPlot = T)
}
\arguments{
\item{dlist}{list of three dataframes : [[1]] Datamatrix, [[2]] SamplesMetadata [[3]] VariableMetadata.}

\item{Samples.grp}{Factor name used for grouping samples (affect plot only)}

\item{Variables.grp}{Factor name used for grouping variables (affect plot only)}

\item{Legend.L}{Logical for drawing legends}

\item{colorL}{Logical for using color or greyscale}

\item{Samp.lab.L}{Logical for drawing Sample labels}

\item{Var.lab.L}{Logical for drawing Variable labels}

\item{palpha}{Transparency for loadings points (between 0 and 1)}

\item{psize}{Loadings points size}

\item{ShowPlot}{Logical to directly draw the plot or return the grobs (faster)}
}
\value{
A list with [1] pca results, [2] plot as grobs.
}
\description{
Perform a PCA analysis on a 3 levels list and create custom plot.
}
\examples{
dlist.pca.old()
}
\keyword{ggplot}
\keyword{pca,}
