% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dlist.pca}
\alias{dlist.pca}
\title{Perform PCA and custom plot}
\usage{
dlist.pca(dlist, group.spl = NULL, group.var = NULL, legendL = F,
  labels = c("scores", "loadings", "both"), pathL = F, ellipseL = F,
  outliersL = F, ShowPlot = T)
}
\arguments{
\item{dlist}{list of three dataframes : [[1]] Datamatrix, [[2]] SamplesMetadata [[3]] VariableMetadata.}

\item{group.spl}{Factor name used for grouping samples (affect plot only)}

\item{group.var}{Factor name used for grouping variables (affect plot only)}

\item{legendL}{Logical for drawing legends}

\item{labels}{String to show labels on scores, loadings or both plots}

\item{pathL}{Loadings points size}

\item{ellipseL}{Logical to show ellipse (95% type t which assumes a multivariate t-distribution)}

\item{outliersL}{Logical to show outliers (below 1 percentile or above 99 percentile)}

\item{ShowPlot}{Logical to draw plot or not}
}
\value{
A list with [1] pca results, [2] plot as grobs.
}
\description{
Perform a PCA analysis on a 3 levels list and create custom plot.
}
\examples{
dlist.pca()
}
\keyword{ggplot}
\keyword{pca,}
