% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dlist.plot.table}
\alias{dlist.plot.table}
\title{dlist plot table}
\usage{
dlist.plot.table(dlist)
}
\arguments{
\item{dlist}{list of three dataframes : [[1]] Datamatrix, [[2]] SamplesMetadata [[3]] VariableMetadata.}

\item{var2names}{List of grouping factor, if blank calculation while be done for eache variables.}

\item{plotL}{Should summary plot be drown (return a grob, use plot or grid::grid.draw to show)}

\item{alpha}{Set the plot transparency}

\item{size}{Set the points sizes}

\item{Class}{Column names for grouping variables}
}
\value{
a data.table in long format (melted) with sample metadata as id variables.
}
\description{
Generate a long table with sample metadata as id variables for plots
}
\examples{
dlist.plot.table()
}
\keyword{summary}
