% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dlist.summary.2}
\alias{dlist.summary.2}
\title{dlist.summary.2}
\usage{
dlist.summary.2(dlist)
}
\arguments{
\item{dlist}{list of three dataframes : [[1]] Datamatrix, [[2]] SamplesMetadata [[3]] VariableMetadata.}
}
\value{
result of the function
}
\description{
Description of the function
}
\examples{
dlist.summary.2()
}
\keyword{x1,}
\keyword{x2,}
\keyword{x3}

