% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dlist.summary}
\alias{dlist.summary}
\title{dlist summary}
\usage{
dlist.summary(dlist, var2names = NULL, plotL = F, alpha = 0.8,
  size = 0.4)
}
\arguments{
\item{dlist}{list of three dataframes : [[1]] Datamatrix, [[2]] SamplesMetadata [[3]] VariableMetadata.}

\item{var2names}{List of grouping factor, if blank calculation while be done for eache variables.}

\item{plotL}{Should summary plot be drown (return a grob, use plot or grid::grid.draw to show)}

\item{alpha}{Set the plot transparency}

\item{size}{Set the points sizes}
}
\value{
a data.table with summary results (length, NA count, Zero count, Perc of zero, mean, median, sum, min, max, SD, CV, IC95 and skewness)
}
\description{
Calculate dlist summary data using data.table structure
}
\examples{
dlist.summary()
}
\keyword{summary}

