% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dlist.transform}
\alias{dlist.transform}
\title{Transform function for dlist}
\usage{
dlist.transform(Data, ...)
}
\arguments{
\item{Data}{a list with 3 dataframes (Datamatrix, SampleMetadata and VariableMetadata)}

\item{...}{Argument passed to dataframe.transform()}
}
\value{
The resulting datamatrix only
}
\description{
Replace zero by the minimum value / 2 and if checked, tranform data with Trans.fun function (by default log2).
Same function as dataframe.transform(), but formatted for direct use on dlist.
}
\examples{
dlist.transform()
}
\keyword{dlist}
\keyword{transform,}

