% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dlist_stat_table1}
\alias{dlist_stat_table1}
\title{Do pairwise statistical tests on dlist}
\usage{
dlist_stat_table1(data, formula, group.by = NULL, ..., debug = F)
}
\arguments{
\item{data}{A dlist}

\item{formula}{The formula to use with ggpubr::compare_means}

\item{group.by}{Grouping variable for ggpubr::compare_means}

\item{...}{Parameters to pass to ggpubr::compare_means}

\item{debug}{Logical to show debugging messages}
}
\value{
result of the function
}
\description{
These function take a dlist as entry and perform pairwise statistical test using ggpubr::compare_means for
tests and multcompView::multcompLetters to annotate test results. It returns a table
}
\examples{
dlist_stat_table1()
}
\keyword{dlist}
\keyword{letters,}
\keyword{significance}
\keyword{statistical}
\keyword{test,}
