% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{find.limits}
\alias{find.limits}
\title{Find scales for plots}
\usage{
find.limits(x,y)
}
\arguments{
\item{x}{Values on the x axis}

\item{y}{Values on the y axis}
}
\value{
4 numerics with max and min value for x and y
}
\description{
Return the limits of two vectors centered on 0 and with a 10 % margin.
}
\examples{
find.limits()
}
\keyword{ggplot}
\keyword{scales,}

