% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{galaxy.save.list}
\alias{galaxy.save.list}
\title{Save file for GALAXY}
\usage{
galaxy.save.list(data.list)
}
\arguments{
\item{x}{3 levels list}

\item{Results.path}{Saving path}

\item{pref}{Prefix to add}
}
\value{
Write 3 csv file on disk
}
\description{
Write 3 csv files from a three levels list to import in GALAXY.
}
\examples{
galaxy.save.list()
}
\keyword{export}
\keyword{galaxy,}
\keyword{list,}

