% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{get_sign}
\alias{get_sign}
\title{get_sign}
\usage{
get_sign(model)
}
\arguments{
\item{a}{model generated with glm (use of $null.deviance, $deviance, df.null and df.residual)}
}
\description{
Get the theoritical significance glm model.
This function come from WinVector (Nina Zumel, http://winvector.github.io/VariableSignal/VariableSignal.html).
}
\examples{
get_sign(model)
}
\keyword{glm,}
\keyword{significance}

