% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{ggplot_opls_vips}
\alias{ggplot_opls_vips}
\title{Plot OPLS results with VIPs}
\usage{
ggplot_opls_vips(data, VIP.thr = 1, xlabsL = T, ShowPlot = T)
}
\arguments{
\item{data}{Result of ropls with opls method}

\item{VIP.thr}{VIP threshold for groups (1 by default)}

\item{ShowPlot}{Logical to directly draw the plot or return the grobs (faster)}
}
\value{
Return a list with Plot and VIPs
}
\description{
Description of the function
}
\examples{
ggplot_opls_vips()
}
\keyword{ggplot,}
\keyword{opls}
