% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{importWorksheets}
\alias{importWorksheets}
\title{Import Excel file to list}
\usage{
importWorksheets(filename)
}
\arguments{
\item{filename}{Path to the excel file to import}
}
\value{
A 3 levels list
}
\description{
Import excel spreadsheet to a 3 levels list. This function doesn't work for big files because of memory issue with Java.
}
\examples{
importWorksheets()
}
\keyword{excel}
\keyword{import,}
\keyword{list,}

