% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{mz.ppm}
\alias{mz.ppm}
\title{Formula to return mz +- ppm}
\usage{
mz.ppm(mass, ppm)
}
\arguments{
\item{ppm}{tolerance in ppm}

\item{mz}{mz to search}

\item{database}{database to search into (must have 'mz' column)}
}
\value{
a vector with mz range given a ppm tolerance
}
\description{
output a vector with lowest and highest values from an mz and ppm
}
\examples{
ppm(125.0000, 5)
}
\keyword{deviation}
\keyword{mz,}
\keyword{ppm,}
