% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{plot.ropls}
\alias{plot.ropls}
\title{Plot ropls results}
\usage{
\method{plot}{ropls}(plot.opls.data, group = NULL)
}
\arguments{
\item{plot.opls.data}{Result of dlist.opls.data function}

\item{group}{Group name for colors}
}
\value{
A grob, use grid::grid.draw() to plot
}
\description{
Plot ropls results using ggplot2
}
\examples{
plot.ropls()
}
\keyword{ggplot}
\keyword{ropls,}

