% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{plot.theme.1}
\alias{plot.theme.1}
\title{Semi-automatic ggplot theme}
\usage{
\method{plot}{theme.1}(Samples.grp = NULL, Variables.grp = NULL,
  limits = NULL, Legend.L = T, colorL = F, labels = list(title = "", x =
  "", y = ""), geom_path = F, labelsL = F)
}
\arguments{
\item{Samples.grp}{Name of the grouping factor for samples}

\item{Variables.grp}{Name of the grouping factor for variables}

\item{limits}{Limits of the axis (returned by find.limits)}

\item{Legend.L}{Logical to draw legend}

\item{colorL}{Logical to use color or grey scale}

\item{labels}{list for titles (title, x, y)}

\item{geom_path}{Logical for drawing path}

\item{labelsL}{Add labels to points}
}
\value{
ggplot list of aestethic
}
\description{
Returns the ploting parameters according to input, with working default for missing ones.
Used for custom ggplot of multivariate results (pca, pls, opls)
}
\examples{
plot.theme.1()
}
\keyword{ggplot}
\keyword{scales,}

