% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{ropls_plot_vips}
\alias{ropls_plot_vips}
\title{Plot VIPs from opls object}
\usage{
ropls_plot_vips(data, n = NULL)
}
\arguments{
\item{data}{opls object returned by ropls::opls()}

\item{n}{number of VIPs from top to show (all if NULL)}
}
\value{
The resulting list of opls function (subsetted if min = TRUE).
}
\description{
Plot the n top Vips in regard to there correlation (positive or negative) along p1.
}
\examples{
dlist.ropls.min()
}
\keyword{ggplot}
\keyword{opls,}
\keyword{vips,}
