% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sacurine.tbl}
\alias{sacurine.tbl}
\title{Sacurine data from ropls}
\format{A dlist : three levels list of tibbles (Datamatrix, SampleMetadata, VariableMetadata)}
\source{
From ropls package, data : sacurine
}
\usage{
data(sacurine.dlist)
}
\description{
Urine samples from 183 human adults were analyzed by liquid chromatography coupled to high-resolution
mass spectrometry (LTQ Orbitrap) in the negative ionization mode. A total of 109 metabolites were
identified or annotated at the MSI level 1 or 2. After retention time alignment with XCMS, peaks were
integrated with Quan Browser. After signal drift and batch effect correction of intensities, each urine
profile was normalized to the osmolality of the sample. Finally, the data were log10 transformed.
}
\examples{
data(sacurine.dlist)
}
\references{
Thevenot E.A., Roux A., Xu Y., Ezan E. and Junot C. (2015). Analysis of the human adult urinary
metabolome variations with age, body mass index and gender by implementing a comprehensive workflow for
 univariate and OPLS statistical analyses. Journal of Proteome Research, DOI: 10.1021/acs.jproteome.5b00354
}
\keyword{datasets}
