% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{scale_pareto}
\alias{scale_pareto}
\title{Pareto scaling}
\usage{
scale_pareto(data, centeringL = F)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe of same dimensions as input
}
\description{
Apply Pareto scaling on a data.table
}
\examples{
scale_pareto()
}
\keyword{scaling}
