% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{SD_batch_Results}
\alias{SD_batch_Results}
\title{SD_batch_Results}
\usage{
SD_batch_Results(Batch.files = Batch.files, xcmsSet_param = list(method =
  "centWave", ppm = 7, peakwidth = c(4, 20), snthresh = 10, prefilter = c(4,
  10000), mzdiff = -0.001, fitgauss = FALSE, nSlaves = 4))
}
\arguments{
\item{Batch.files}{List returned by SD_mass_batch with files for each batch.}

\item{xcmsSet_param}{xcmsSet parameters (can be any parameters formated like this : c(method="centWave", "ppm=7))}
}
\description{
This function separate .mzXML file list by batch. Useful when you use directory for your classes
and want to do separate xcms analysis by another grouping factor like batch sequences.
}
\keyword{orbitrap}
\keyword{xcms,}

