% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{SD_files_class}
\alias{SD_files_class}
\title{SD_files_class}
\usage{
SD_files_class(dir_path)
}
\arguments{
\item{dir_path}{Origin directory pathway.}
}
\description{
This function get multiple files path and name, and then parse the information into a dataframe,
making files grouping by directory path easy.
}
\examples{
SD_files_class(dir_path = "./")
}
\keyword{Files,}
\keyword{info}

