% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{SD_mass_batch}
\alias{SD_mass_batch}
\title{SD_mass_batch}
\usage{
SD_mass_batch(Files.dir = "./", Batch.list = Batch.list)
}
\arguments{
\item{Files.dir}{Path to your results from workdirectory.}

\item{Batch.list}{A dataframe with rownames is files names (without extension) and first column is batch number (or other factor).}
}
\description{
This function separate .mzXML file list by batch. Useful when you use directory for your classes
and want to do separate xcms analysis by another grouping factor like batch sequences.
}
\keyword{orbitrap}
\keyword{xcms,}

