% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{SD_pca}
\alias{SD_pca}
\title{SD_pca}
\usage{
SD_pca(Data_pca, Folder_name = "Default", ropls_param = list(predI = 2,
  plotL = F), factor_group = NULL)
}
\arguments{
\item{ropls_param}{Parameters to pass to ropls::opls function}

\item{Data}{list with [[1]] Datamatrix [[2]] samples metadata}

\item{Results.path.root}{Name for the results folder}
}
\description{
Perform pca according to ropls method and save results in a folder.
You can choose to color samples by factor or value contained in a sample metadata table.
}
\examples{
SD_pca(Data, Results.path.root = "Default", ropls_param = list(predI = 2, plotL = F), factor_group = NULL)
}
\keyword{pca}

