% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{SD_subset_zero}
\alias{SD_subset_zero}
\title{SD_subset_zero}
\usage{
SD_subset_zero(x, Z.Seuil = 80)
}
\arguments{
\item{x}{Dataframe (a datamatrix) with only numeric values}

\item{Z.Seuil}{Percentage threshold of zero values to accept for each variable}
}
\description{
Subset a dataframe containing numeric value by selecting column which have less than
Z.Seuil percents of zero values.
}
\examples{
SD_subset_zero(Data[[1]], Z.Seuil = 50)
}
\keyword{pca}

