% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{size}
\alias{size}
\title{Get size of object or list objects}
\usage{
size(data)
}
\arguments{
\item{data}{Any variable}
}
\value{
Object size
}
\description{
Return the size in Mb of an object or list levels
}
\examples{
size()
}
\keyword{x1,}
\keyword{x2,}
\keyword{x3}
