% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{splitdf}
\alias{splitdf}
\title{Split a dataframe}
\usage{
splitdf(cars, p = 0.5, seed = 95687)
}
\arguments{
\item{dataframe}{a dataframe to subset.}

\item{p}{the proportion of data to keep in trainset}

\item{seed}{the seed number for repetability (same seed will generate same subset for a given dataframe).}
}
\value{
A list with [1] trainset dataframe, [2] testset dataframe.
}
\description{
Split a dataframe in two.
}
\examples{
splitdf()
}
\keyword{dataframe,}
\keyword{subset}
\keyword{training,}
