% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{table_summary}
\alias{table_summary}
\title{Summary on table}
\usage{
table_summary(data, val.name, group.by)
}
\arguments{
\item{data}{A dlist}

\item{val.name}{Column name of value to summarize}

\item{group.by}{Column(s) name(s) of grouping factors}
}
\value{
Return a data.table with summary
}
\description{
Perform classic statistical indices on a table
}
\examples{
stat_table()
}
\keyword{summary}
