% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{test}
\alias{test}
\title{TITLE}
\usage{
test(x)
}
\arguments{
\item{x}{Argument description}
}
\value{
result of the function
}
\description{
Description of the function
}
\examples{
test()
}
\keyword{x1,}
\keyword{x2,}
\keyword{x3}

