% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{testing_function_01}
\alias{testing_function_01}
\title{First function}
\usage{
testing_function_01(x)
}
\arguments{
\item{x}{Argument for [...]}
}
\description{
This function is just a test for R package creation.
}
\examples{
testing_function_01(1)
}
\keyword{Test}

