% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{to.data.table}
\alias{to.data.table}
\title{Check dlist table format and convert to data.table}
\usage{
to.data.table(dlist, rownamesL = F)
}
\arguments{
\item{dlist}{Three levels list with [[1]] Datamatrix, [[2]] SamplesMetadata, [[3]] VariableMetadata.}

\item{rownamesL}{Logical to specify wether the input dlist use data.frame rownames to store rowID.}
}
\value{
The dlist with data converted to dataframe. Also print a check status of class used in the
dlist
}
\description{
Check the format of 3 levels list tables print a class summary and convert to
data.table if not. User can specify if there is rownames (if data.frame were used in entry) so
the function can place them as first variable (since data.table does not use rownames logic).
}
\examples{
to.data.table()
}
\keyword{check,}
\keyword{data.table}
\keyword{list,}

