% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{transf_glog}
\alias{transf_glog}
\title{glog transformation}
\usage{
transf_glog(data, a = 1)
}
\arguments{
\item{data}{a dataframe}

\item{a}{scaling factor}
}
\value{
a dataframe of same dimensions as input
}
\description{
Apply glog transformation, which behave better than log transformation with zero values
}
\examples{
transf_glog()
}
\keyword{transform}
