% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{waters_txt}
\alias{waters_txt}
\title{Read and format MassLynx text file}
\usage{
waters_txt(file)
}
\arguments{
\item{file}{Path to the txt file}
}
\value{
a data.table with
}
\description{
Read summary exported from MassLynx as txt files and format the content to output one table
with an ID column.
}
\examples{
waters_txt()
}
\keyword{Format}
\keyword{Read,}
