% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{write.csv3}
\alias{write.csv3}
\title{Write CSV for Excel Fr}
\usage{
write.csv3(data, file = "./Results/Table_x.csv")
}
\arguments{
\item{data}{Dataframe to save}

\item{file}{File path and name}
}
\value{
Write a csv file
}
\description{
Write a csv directly sees as a table by excel.
}
\examples{
write.csv3()
}
\keyword{dataframe,}
\keyword{excel,}
\keyword{save}
