% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{xcms_orbi_GRT}
\alias{xcms_orbi_GRT}
\title{xcms_orbi_GRT}
\usage{
xcms_orbi_GRT(xcms_set_obj, Results.dir.name="Default", bw_param=c(25, 10, 0.7), mzwid_param=0.005, minfrac_param=0.25, profStep_param=0.8)
xcms_orbi_GRT()
}
\arguments{
\item{File_list}{File list to pass to xcmsSet method = 'centWave', ppm=7, peakwidth=c(4,20), snthresh=10, prefilter=c(4,1000), mzdiff=-0.001, fitgauss=F, nSlaves = 4}

\item{Results.dir.name}{Name of the subfolder to store results}

\item{bw_param}{Vector for bw settings to use in 1, 2 and 3 iteration : c(1, 2, 3)}

\item{mzwid_param}{mzwid parameter to use}

\item{minfrac_param}{minfrac parameter to use}

\item{profStep_param}{profStep parameter to use}

\item{Sample.Metadata}{Dataframe with samples names and metadata}

\item{Grouping.factor}{Col.number of Sample.Metadata to use for ACP colors (can be a vector)}
}
\description{
This function simplify workflow from peak grouping and retention time correction
to PCA analysis. Function will perform group -> retcor -> group -> retcor -> group
and return EIC for standards across all samples and extract standards from peaks table,
giving also ppm deviation for each std for diagnostic.
}
\keyword{orbitrap}
\keyword{xcms,}

