% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{xcms_orbi_Results}
\alias{xcms_orbi_Results}
\title{xcms_orbi_Results}
\usage{
xcms_orbi_GRT(xcms_set_obj, Results.dir.name="Default", bw_param=c(25, 10, 0.7), mzwid_param=0.005, minfrac_param=0.25, profStep_param=0.8)
xcms_orbi_GRT()
}
\arguments{
\item{filled_peak_object}{An xcmsSet object with filled peaks}

\item{STDs_mass}{Vector of STDs exact mass}

\item{STDs_ppm}{Deviation for STDs in ppm (try to increase if your STDs are not found)}

\item{perform_PCA}{Logical to do PCA analysis or not}

\item{Sample.Metadata}{Dataframe with samples metadata}

\item{PCA_group}{vector with samples metadata column to use for groups in ACP}
}
\description{
This function generate peak_table, STDs EIC across samples, PCA (optional)
and return a list with [1] datamatrix, [2] sample.metadata and [3] variable.metadata.
}
\keyword{orbitrap}
\keyword{xcms,}
