% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{xcmsSet.Result.List}
\alias{xcmsSet.Result.List}
\title{3 Levels list from xcmsSet Object}
\usage{
xcmsSet.Result.List(x)
}
\arguments{
\item{x}{xcmsSet Object}
}
\value{
A dlist : 3 levels list (Datamatrix, SampleMetadata, VariableMetadata)
}
\description{
Create a 3 levels list from the result of xcms integration.
}
\examples{
xcmsSet.Result.List()
}
\keyword{3levels.list,}
\keyword{xcmsSet}

