% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{check.list.format}
\alias{check.list.format}
\title{Check format of 3 levels list}
\usage{
check.list.format(data)
}
\arguments{
\item{data}{Three levels list with [[1]] Datamatrix, [[2]] SamplesMetadata, [[3]] VariableMetadata.}
}
\description{
Check the format of 3 levels list.
}
\examples{
check.list.format()
}
\keyword{check}
\keyword{list,}

