% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{check.list.format}
\alias{check.list.format}
\title{Check format of dlist}
\usage{
check.list.format(dlist, to.data.table.L = T)
}
\arguments{
\item{dlist}{Three levels list with [[1]] Datamatrix, [[2]] SamplesMetadata, [[3]] VariableMetadata.}

\item{to.data.table}{Logical to convert data.frame's dlist to data.table with to.data.table function}
}
\value{
Print result of check as character and return the dlist (or converted dlist) if format is ok
}
\description{
Check the format of 3 levels list.
}
\details{
Chech if the list elements are data.table or data.frame, id rownames [[1]] and [[2]] are identical (for
data.frame only, use first column as row id with data.table), and if colnames [[1]] are identical to
rownames [[3]]. Check also dataframes dimension to check consistency and propose to transform data.frame
to data.table.
}
\examples{
check.list.format()
}
\keyword{check}
\keyword{list,}

