% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{data.subset}
\alias{data.subset}
\title{data.subset}
\usage{
data.subset(data, Var.sel = NULL, Samples.sel = NULL)
}
\arguments{
\item{data}{list of three dataframes : [[1]] Datamatrix, [[2]] SamplesMetadata [[3]] VariableMetadata.}

\item{Var.sel}{vector of variable to subset (rownames) from [[3]]}

\item{Samples.sel}{vector of samples to subset (rownames) from [[2]]}
}
\description{
Subset a list of three dataframes : [[1]] Datamatrix, [[2]] SamplesMetadata [[3]] VariableMetadata.
[[1]] samples.ID x variables.ID
[[2]] samples.ID x sample.metadata (same order than rows in [[1]])
[[3]] variables.ID x variable.metadata (same order than columns in [[1]])
}
\examples{
data.subset(data, Var.sel = c(1,4,5,8), Samples.sel = c(1,8,9,15))
}
\keyword{list}
\keyword{subset,}

