% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{data.ttest}
\alias{data.ttest}
\title{3 levels t-test}
\usage{
data.ttest(dlist, group)
}
\arguments{
\item{dlist}{list of three dataframes : [[1]] Datamatrix, [[2]] SamplesMetadata [[3]] VariableMetadata.}

\item{group}{The grouping factor}
}
\value{
A dataframe with p.valus, x and y estimates, delta mean and CV
}
\description{
Perform t-test on a 3 levels list between groups
}
\examples{
data.ttest()
}
\keyword{t-test}
