% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{densplot}
\alias{densplot}
\title{2 dimension plot with density}
\usage{
densplot(Data, x, y, fill = "grey", color = "black",
  labels = list("title", "x", "y"))
}
\arguments{
\item{Data}{Long format data to plot}

\item{x}{column with x data}

\item{y}{column with y data}

\item{fill}{filling color or factor}

\item{color}{string for point color}

\item{labels}{List for labels with (title, x, y)}
}
\value{
a ggplot
}
\description{
Draw a 2 dimensions plot with density graph on each axis
}
\examples{
densplot()
}
\keyword{x1,}
\keyword{x2,}
\keyword{x3}

