% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dlist.opls}
\alias{dlist.opls}
\title{Perform OPLS and custom plots}
\usage{
dlist.opls(dlist, Opls.y = NA, ..., Samples.grp = NULL,
  Variables.grp = NULL, Legend.L = T, colorL = F, VIP.thr = 1,
  LabelsL = F, ShowPlot = T, xlabsL = T)
}
\arguments{
\item{dlist}{list of three dataframes : [[1]] Datamatrix, [[2]] SamplesMetadata [[3]] VariableMetadata.}

\item{Opls.y}{Name of the grouping factor used in OPLS}

\item{...}{Arguments to pass to opls function}

\item{Samples.grp}{Vector of samples grouping factor (for plot color only)}

\item{Variables.grp}{Vector of variable grouping factor (for plot color only)}

\item{Legend.L}{Logical to draw legends}

\item{colorL}{Logical to use colors instead of black & white}

\item{LabelsL}{Logical to show labels on plot}

\item{ShowPlot}{Logical to draw plots in graphic device}

\item{xlabsL}{Logical to show VIPs labels}
}
\value{
A list with (OPLS, Plot, VIPS)
}
\description{
Perform an OPLS analysis using "ropls" package and plot results if asked (scores, loadings, VIPs & summary).
}
\examples{
dlist.opls()
}
\keyword{ggplot}
\keyword{opls,}
