% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{dlist.summary}
\alias{dlist.summary}
\title{Overview of 3 level datas}
\usage{
dlist.summary(dlist, var2.names = NULL, IC = F, plotL = F, alpha = 0.2,
  labels = list(title = "", x = "", y = ""), x.labL = F)
}
\arguments{
\item{dlist}{list of three dataframes : [[1]] Datamatrix, [[2]] SamplesMetadata [[3]] VariableMetadata.}

\item{var2.names}{vector of grouping factors names}

\item{IC}{Logical for IC95 calculation (could be slow if true)}

\item{plotL}{Logical for generating plot}

\item{alpha}{Transparency value of graphs}

\item{labels}{list of graph labels (title, x, y)}

\item{x.labL}{Logical for showing point labels}
}
\value{
Return a list with the calculations results as a data frame and the plot if asked.
}
\description{
Perform a quick analysis of a 3 levels list by showing for each variables : the average value,
the Coefficient of Variation, the skew and the percentage of zero values. Calculations is done by
grouping factors for each variables and returned in a list with the plot.
}
\examples{
dlist.summary()
}
\keyword{graph}
\keyword{summary,}

