% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{g_legend}
\alias{g_legend}
\title{Get legend of a ggplot}
\usage{
galaxy.save.list(data.list)
}
\arguments{
\item{a.gplot}{A ggplot}
}
\value{
Write 3 csv file on disk
}
\description{
Get the legend of a ggplot to draw it externaly & draw it as grob.
}
\examples{
g_legend()
}
\keyword{ggplot}
\keyword{legend,}
