% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{importWorksheets.2}
\alias{importWorksheets.2}
\title{Import Excel file to list}
\usage{
importWorksheets.2(Data.path, fcolL = T)
}
\arguments{
\item{Data.path}{Path to the excel file to import}

\item{fcolL}{Logical to specify if th first column should be use for rownames (must contains unique names)}
}
\value{
A list of dataframe
}
\description{
Import each sheet of an excel file to a list. This function use the readxl package, which seems
to handle large files.
}
\examples{
importWorksheets.2()
}
\keyword{excel}
\keyword{import,}
\keyword{list,}

