% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.1_functions.R
\name{SD_pca}
\alias{SD_pca}
\title{SD_pca}
\usage{
SD_pca(Data, Results.path.root = "Default", ropls_param = list(predI = 2,
  plotL = F), factor_color = NULL)
}
\arguments{
\item{Data}{Dataframe (rownames = samples ; colnames = measures)}

\item{Results.path.root}{Name for the results folder}

\item{ropls_param}{Parameters to pass to ropls::opls function}

\item{factor_color}{Factor list to use for grouping}
}
\description{
Perform pca according to ropls method and save results in folder.
}
\examples{
SD_pca(Data, Results.path.root = "Default", ropls_param = list(predI = 2, plotL = F))
}
\keyword{pca}

